refactor: Add `ItemMarkdown` component
We need to reuse the functionality to insert the multi-informant badge
above markdown content, and the "Optional" text below the markdown,
inside `RequestHealthRecordDataItem` and `ActivityCardItem` components.
Created a new component `ItemMarkdown` to be reused in both places.

Slightly refactored `RequestHealthRecordDataItem` to include the
external link inside the unprocessed markdown so that the "Optional"
label is displayed below it.
