+++ src/shared/ui/CardItem/index.tsx
@@ -1,11 +1,7 @@
-import { PropsWithChildren, useContext, useMemo } from 'react';
+import { PropsWithChildren } from 'react';
 
-import { TargetSubjectLine } from './TargetSubjectLine';
-
-import { SurveyContext } from '~/features/PassSurvey';
-import { Theme } from '~/shared/constants';
-import { Box, Markdown, Text } from '~/shared/ui';
-import { insertAfterMedia, useCustomMediaQuery, useCustomTranslation } from '~/shared/utils';
+import { Box, ItemMarkdown } from '~/shared/ui';
+import { useCustomMediaQuery } from '~/shared/utils';
 
 interface CardItemProps extends PropsWithChildren {
   watermark?: string;
@@ -18,18 +14,6 @@ interface CardItemProps extends PropsWithChildren {
 export const CardItem = ({ children, markdown, isOptional, testId }: CardItemProps) => {
   const { greaterThanSM } = useCustomMediaQuery();
 
-  const { t } = useCustomTranslation();
-
-  const context = useContext(SurveyContext);
-
-  const processedMarkdown = useMemo(() => {
-    if (markdown !== null && markdown !== undefined) {
-      if (!context.targetSubject) return markdown;
-      return insertAfterMedia(markdown, '<div id="target-subject"></div>');
-    }
-    return markdown;
-  }, [markdown, context.targetSubject]);
-
   return (
     <Box
       data-testid={testId || 'active-item'}
@@ -40,33 +24,7 @@ export const CardItem = ({ children, markdown, isOptional, testId }: CardItemPro
       gap="48px"
       sx={{ fontWeight: '400', fontSize: '18px', lineHeight: '28px' }}
     >
-      {processedMarkdown ? (
-        <Box>
-          <Markdown
-            markdown={processedMarkdown}
-            components={{
-              div: (props) =>
-                props.id === 'target-subject' ? (
-                  <TargetSubjectLine subject={context.targetSubject} />
-                ) : (
-                  <div {...props} />
-                ),
-            }}
-          />
-          {isOptional && (
-            <Text
-              variant="body1"
-              color={Theme.colors.light.outline}
-              testid="optional-item-label"
-              fontWeight="400"
-              fontSize="18px"
-              lineHeight="28px"
-            >
-              {`(${t('optional')})`}
-            </Text>
-          )}
-        </Box>
-      ) : null}
+      <ItemMarkdown markdown={markdown ?? ''} isOptional={isOptional} />
       <Box>{children}</Box>
     </Box>
   );
